Rename DailyRewardListener->PlayerJoinListener; add store-on-join config; register new listener
