+++ src/main/resources/config.yml
@@ -23,6 +23,12 @@ daily-reward:
   sound: ENTITY_EXPERIENCE_ORB_PICKUP
   debug: false
 
+# When enabled, the storage provider will be asked to create/store a record
+# for each player when they join the server. This helps ensure player data
+# files/rows exist for admin reporting tools. Default: false
+store-on-join:
+  enabled: false
+
 # Multi-currency support (optional)
 multi-currency:
   enabled: false
